requirements.txt
Flask~=3.1.0
protobuf~=5.29.4
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.166.0
pytz~=2025.2
langchain-google-genai~=2.1.2
langchain-core~=0.3.50
pydantic~=2.11.1
python-telegram-bot~=22.0
requests~=2.32.3
google~=3.0.0
twilio~=9.5.2